feat(VaultUnlocker): Consider an empty vault as unlocked
We consider a user doesn't use the password manager feature as long as
he doesn't have any cipher. In that case, we don't want him to unlock
his empty unused vault. So we consider that an empty vault as always
unlocked. When the user uses the Cozy web browser extension for the
first time, his existing accounts will be added to its vault, then we
will consider that he is using the vault.