feat: Check if extension has been installed before ask for unlock
We figured than only checking if some ciphers exist is not a strong
enough condition to know if we should ask the user to unlock the vault
or not.

To resolve our problems, the stack now adds a
`extension_installed` attribute on the `io.cozy.settings.bitwarden`
document when an extension has been connected for the first time to the
instance. See https://github.com/cozy/cozy-stack/pull/2315.

So we now check this attribute. The flow is now:

* Check if there is some cipher, is yes => the user should unlock
* If there is no cipher, check if the extension has been installed, if
  yes => the user should unlock
* Otherwise, skip the unlock step