feat: Add import
We can't use the ImportService's `import` method because it does a call
to a `/ciphers/import`, which is not yet implemented by the stack. So I
reimplemented a basic import feature that only handles the ciphers. It
only uses the jslib's import formats and parsers so we cover all
possible formats.

For each imported cipher, it looks for an existing one by
{username,password,uri} tuple. If it finds an existing cipher, it adds
the URIs that are in the imported cipher but not in the existing one to
it. If it doesn't find an existing cipher, it creates a new one with the
imported cipher data.