refactor: Separate cipher matching from cipher saving
Before, the matching of a cipher was done just before saving.
For batch import, we need to be able to have a matching phase for all
ciphers and then a saving phase where we send all ciphers. Here we
refactor the code to make the matching phase more apparent (mapping all
ciphers into prepareCipherToImport) and then we save them all. In a
future commit, the map to saveCipher will be replaced by a batch save.