feat: Move VaultUnlock components from harvest
Moved code that was originally in Harvest so that all code managing the
vault is in cozy-keys-lib
Additionally, an additional check in showUnlockForm has been added to
check if the extension has been installed. This way, the user of the
lib, does not need to do the check. If the extension is not installed
or if the vault already is unlocked, onUnlock is called immediately.