feat: Use isPassContext flag to check if extension is installed
In a pass context, we want to check if a Pass client is installed, e.g.
the browser extension, the mobile app or the web cozy-pass client. This
check is useful to know whether or not the unlock vault form should be
displayed to the user.

But in other context, such as Drive client-side encryption, this check
is not required at all.