fix: Remove broken UnlockForm export
Since cc8f07932ca400afce1d82f3d73c27f6bfe28074, the UnlockForm is
exported. However, this was wrongly exported as it is default exported
in the component. Thus, this cannot work and triggers a warning in apps
using the lib such as Drive: `export 'UnlockForm' was not found`.
Consequently, we remove this useless export